+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Feature: click "xx quests loaded" to check for updates to quest data from server -alydev
 - Feature: Open Editor feature -alydev
   - This currently opens the quest path JSON in your default text editor, or generates a new quest path JSON for the current quest. A GUI quest editor will be included Soon :tm:
   - The Edit button now has a tooltip showing how it works. If quests stop working, right click this button and delete all quest paths in that folder first to see if that fixes it.